Seperated Store and Routes for Multi-Team Workflow Synergy
