+++ src/App.tsx
@@ -4,14 +4,14 @@ import 'react-toastify/dist/ReactToastify.css';
 import './App.scss';
 import { store } from './Store';
 import AppNav from './components/nav/nav.component';
-import { BrowserRouter, Route, Switch } from 'react-router-dom';
-import { HomeComponent } from './components/home/home.component';
+import { BrowserRouter,  Switch } from 'react-router-dom';
+
 import { Provider } from 'react-redux';
-import clickerContainer from './components/clicker/clicker.container';
-import ManageComponent from './components/manage/manage.container';
-import  ProtectedRoute  from './components/protected-route.component/protected-route.component';
-import  LoginComponent  from './components/login/login.component';
+
 import { ToastContainer, toast } from 'react-toastify';
+import { ManagementRoutes } from './routes/Management.routes';
+import { InterviewRoutes } from './routes/Interview.routes';
+import { SurveyRoutes } from './routes/Survey.routes';
 
 class App extends React.Component {
   public render() {
@@ -22,11 +22,9 @@ class App extends React.Component {
             <AppNav />
             <div id="main-content-container">
               <Switch>
-                <Route path="/login" component={LoginComponent} />
-                <Route path="/home" component={HomeComponent} />
-                <Route path="/clicker" component={clickerContainer} />
-                <ProtectedRoute allowedRoles={['admin', 'staging-manager', 'trainer']} path="/manage/:manage" component={ManageComponent} />
-                <Route component={HomeComponent} />
+                  <ManagementRoutes/>
+                  <InterviewRoutes/>
+                  <SurveyRoutes/>
               </Switch>
             </div>
             <ToastContainer autoClose={2000} position={toast.POSITION.BOTTOM_RIGHT}/>
