+++ src/components/clicker/clicker.container.tsx
@@ -4,7 +4,7 @@ import { increment } from '../../actions/clicker/clicker.actions'
 import { ClickerComponent } from './clicker.component';
 
 
-const mapStateToProps = (state: IState) => (state.clicker);
+const mapStateToProps = (state: IState) => (state.managementState.clicker);
 
 const mapDispatchToProps = {
   increment,
