+++ src/components/manage/manage.container.tsx
@@ -1,7 +1,8 @@
 import { manageGetUsersByGroup } from '../../actions/manage-users/manage-users.actions';
 import * as createUserActions from '../../actions/create-user/create-user.actions';
 import * as createCohortActions from '../../actions/create-cohort/create-cohort.actions';
-import { IState, IManageUsersState } from '../../reducers';
+import { IState, } from '../../reducers';
+import {IManageUsersState } from '../../reducers/management'
 import { connect } from 'react-redux';
 import { RouteComponentProps } from 'react-router';
 import { ManageComponenet } from './manage.component';
@@ -14,7 +15,7 @@ export interface IManageComponentProps extends RouteComponentProps<{manage: stri
 }
 
 const mapStateToProps = (state:IState) => ({
-  manageUsers: state.manageUsers
+  manageUsers: state.managementState.manageUsers
 });
 
 const mapDispatchToProps = {
