+++ src/reducers/interview/index.ts
@@ -1,0 +1,9 @@
+import { combineReducers } from "redux";
+
+export interface IInterviewState {
+    
+  }
+
+  export const interviewState = combineReducers<IInterviewState>({
+    
+  })
