+++ src/reducers/survey/index.ts
@@ -1,0 +1,9 @@
+import { combineReducers } from "redux";
+
+export interface ISurveyState {
+    
+  }
+
+  export const surveyState = combineReducers<ISurveyState>({
+    
+  })
