+++ src/routes/Interview.routes.tsx
@@ -1,0 +1,17 @@
+import React from 'react'
+
+
+export class InterviewRoutes extends React.Component<any, any> {
+    constructor(props){
+        super(props)
+    }
+
+    render() {
+
+        return (
+            <>
+            
+            </>
+        )
+    }
+}
