+++ src/routes/Survey.routes.tsx
@@ -1,0 +1,17 @@
+import React from 'react'
+
+
+export class SurveyRoutes extends React.Component<any, any> {
+    constructor(props){
+        super(props)
+    }
+
+    render() {
+
+        return (
+            <>
+            
+            </>
+        )
+    }
+}
